Update task #15

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -19,17 +19,17 @@ int main(void)
  */
 void fibonacci(void)
 {
-	int n1 = 1, n2 = 2, n3 = 0, temp;
+	size_t n1 = 1, n2 = 2, n3 = 0, temp;
 	int i = 0;
 
-	printf("%i, %i, ", n1, n2);
+	printf("%li, %li, ", n1, n2);
 
 	for (; i < 97; i++)
 	{
 		n3 = n2 + n1;
 		if (i)
 			printf(", ");
-		printf("%i", n3);
+		printf("%li", n3);
 		temp = n3;
 		n1 = n2;
 		n2 = temp;
